+++ OfficeDevPnP.Core/OfficeDevPnP.Core.Tests/App.config
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <appSettings>
+    <add key="SPOTenantUrl" value="https://bertonline-admin.sharepoint.com" />
+    <add key="SPODevSiteUrl" value="https://bertonline.sharepoint.com/sites/dev4" />
+    <add key="SPOUserName" value="bert.jansen@bertonline.onmicrosoft.com" />
+    <add key="SPOPassword" value="MSFT7ever" />
+    <add key="Realm" value="d8623c9e-30c7-473a-83bc-d907df44a26e" />
+    <add key="AppId" value="a4cdf20c-3385-4664-8302-5eab57ee6f14" />
+    <add key="AppSecret" value="I0yAoALDlq96Z6ABpCj4Nfws/huxEiYPb7Id/TvinKE=" />
+  </appSettings>
+</configuration>
