+++ OfficeDevPnP.Core/OfficeDevPnP.Core/OfficeDevPnP.Core.csproj
@@ -253,6 +253,12 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>CoreResources.resx</DependentUpon>
     </Compile>
+    <Compile Include="Framework\TimerJobs\Enums\AuthenticationType.cs" />
+    <Compile Include="Framework\TimerJobs\TimerJob.cs" />
+    <Compile Include="Framework\TimerJobs\TimerJobRun.cs" />
+    <Compile Include="Framework\TimerJobs\TimerJobRunEventArgs.cs" />
+    <Compile Include="Utilities\Log.cs" />
+    <Compile Include="Framework\TimerJobs\Utilities\SiteEnumeration.cs" />
     <Compile Include="IdentityModel\TokenProviders\ADFS\BaseProvider.cs" />
     <Compile Include="IdentityModel\TokenProviders\ADFS\UsernameMixed.cs" />
     <Compile Include="IdentityModel\WSTrustBindings\UserNameWSTrustBinding.cs" />
