+++ OfficeDevPnP.Core/OfficeDevPnP.Core/Utilities/Log.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OfficeDevPnP.Framework.TimerJob.Utilities
+namespace OfficeDevPnP.Core.Utilities
 {
     /// <summary>
     /// Logging class
